workloads/requirements.txt
pandas
absl-py
numpy